requirements_brief_django2.0.10.txt
future==0.17.1
python-dateutil==2.7.5
python-magic==0.4.15
python-magic-bin==0.4.14
Cython==0.29.3
decorator==4.3.2
Send2Trash==1.5.0
simplegeneric==0.8.1
termcolor==1.1.0
testpath==0.4.2
toolz==0.9.0
h5py==2.7.1
httplib2==0.12.0
MarkupSafe==1.1.0
Pygments==2.3.1
pygame==1.9.4
pyparsing==2.3.1
pyzmq==17.1.2
mysqlclient==1.4.1
requests==2.21.0
Django==2.0.10
django-reversion==3.0.3
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-formtools==2.1
XlsxWriter==1.1.2
django-import-export==1.0.0
Markdown==3.0.1
Pillow==5.4.1
django-markdownx==2.0.28
django-pure-pagination==0.3.0
django-simple-captcha==0.5.6
# restful framework
coreapi==2.3.3
markdown==3.0.1
djangorestframework==3.9.1
django-filter==2.1.0
django-crispy-forms
django-guardian==1.5.0
